new confits

+++ python/dolma/cli/deduper.py
@@ -1,14 +1,14 @@
 from contextlib import ExitStack
 from dataclasses import dataclass
+from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 import smart_open
-from cached_path import cached_path
 from omegaconf import OmegaConf as om
 
 from dolma import deduper
 from dolma.cli import BaseCli, field, print_config
-from dolma.cli.shared import WorkDirConfig, make_temp_bloom, make_workdirs
+from dolma.cli.shared import WorkDirConfig, get_path_to_temp_file, make_workdirs
 from dolma.core.errors import DolmaConfigError
 from dolma.core.loggers import get_logger
 from dolma.core.paths import glob_path
@@ -29,7 +29,7 @@ class DocumentDedupeConfig:
 class BloomFilterConfig:
     file: str = field(help="Path where to read/write the bloom filter file to/from. Required.")
     size_in_bytes: int = field(
-        default=-1,
+        default=0,
         help=(
             "Size of the bloom filter in bytes. Either this value is provided, or both estimated_doc_count "
             "and desired_false_positive_rate."
@@ -37,14 +37,14 @@ class BloomFilterConfig:
     )
     read_only: bool = field(help="If true, the bloom filter will be read from the file and not updated. Required.")
     estimated_doc_count: int = field(
-        default=-1,
+        default=0,
         help=(
             "Estimated number of documents to be added to the bloom filter. Either this value is provided, "
             "or both size_in_bytes and desired_false_positive_rate."
         ),
     )
     desired_false_positive_rate: float = field(
-        default=-1.0,
+        default=0,
         help=(
             "Desired false positive rate. Either this value is provided, or both size_in_bytes and "
             "estimated_doc_count."
@@ -119,15 +119,22 @@ class DeduperCli(BaseCli):
                 # but raise an error if no documents are found for all paths
                 raise DolmaConfigError(f"No documents found for the paths {parsed_config.documents}.")
 
-            try:
-                local_bloom_file = cached_path(parsed_config.bloom_filter.file)
-            except FileNotFoundError as ex:
-                if parsed_config.bloom_filter.read_only:
-                    raise ex
-                local_bloom_file = stack.enter_context(make_temp_bloom())
+            # The rust deduper does not work with remote files, so we need to download the bloom filter
+            # if it is not local. If the remote file does not exists, and the bloom filter is read-only,
+            # we raise an error.
+            path_is_local = (local_bloom_file := Path(parsed_config.bloom_filter.file)).exists()
+            if not path_is_local:
+                local_bloom_file = stack.enter_context(get_path_to_temp_file())
+                try:
+                    with smart_open.open(parsed_config.bloom_filter.file, "rb") as f:
+                        contents: bytes = f.read()  # pyright: ignore
+                    local_bloom_file.write_bytes(contents)
+                except (FileNotFoundError, OSError) as ex:
+                    if parsed_config.bloom_filter.read_only:
+                        raise ex
 
             dict_config["bloom_filter"] = {
-                "file": local_bloom_file,
+                "file": str(local_bloom_file),
                 "read_only": parsed_config.bloom_filter.read_only,
                 "size_in_bytes": parsed_config.bloom_filter.size_in_bytes,
                 "estimated_doc_count": parsed_config.bloom_filter.estimated_doc_count,
@@ -142,10 +149,6 @@ class DeduperCli(BaseCli):
                     "Either bloom_filter.size_in_bytes or bloom_filter.estimated_doc_count and "
                     "bloom_filter.desired_false_positive_rate must be specified"
                 )
-
-            for k in list(dict_config["bloom_filter"].keys()):
-                if dict_config["bloom_filter"][k] == -1:
-                    del dict_config["bloom_filter"][k]
 
             dict_config["work_dir"] = {"input": work_dirs.input, "output": work_dirs.output}
             dict_config["processes"] = parsed_config.processes
@@ -163,8 +166,8 @@ class DeduperCli(BaseCli):
             deduper(dict_config)
 
             # upload to remote file if necessary
-            if not parsed_config.bloom_filter.read_only:
-                logger.info("Saving bloom filter to %s", parsed_config.bloom_filter.file)
+            if not parsed_config.bloom_filter.read_only and not path_is_local:
+                print(f"Pushing Bloom filter to {parsed_config.bloom_filter.file}")
                 local = stack.enter_context(smart_open.open(local_bloom_file, "rb"))
                 remote = stack.enter_context(smart_open.open(parsed_config.bloom_filter.file, "wb"))
                 remote.write(local.read())
